refactor: baseentity 경로 변경
