+++ src/main/java/com/springboot/monew/newsarticles/entity/ArticleInterest.java
@@ -1,6 +1,6 @@
 package com.springboot.monew.newsarticles.entity;
 
-import com.springboot.monew.base.BaseEntity;
+import com.springboot.monew.common.entity.BaseEntity;
 import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -32,18 +32,18 @@ public class ArticleInterest extends BaseEntity {
     )
     private NewsArticle newsArticle;
 
-    @ManyToOne(fetch = FetchType.LAZY, optional = false)
-    @JoinColumn(
-            name = "interest_id",
-            nullable = false,
-            foreignKey = @ForeignKey(name = "FK_ARTICLE_INTERESTS_INTEREST")
-    )
-    private Interests interest;
-
-    public ArticleInterest(NewsArticle newsArticles, Interests interest) {
-        this.newsArticle = newsArticles;
-        this.interest = interest;
-    }
+//    @ManyToOne(fetch = FetchType.LAZY, optional = false)
+//    @JoinColumn(
+//            name = "interest_id",
+//            nullable = false,
+//            foreignKey = @ForeignKey(name = "FK_ARTICLE_INTERESTS_INTEREST")
+//    )
+//    private Interests interest;
+//
+//    public ArticleInterest(NewsArticle newsArticles, Interests interest) {
+//        this.newsArticle = newsArticles;
+//        this.interest = interest;
+//    }
